+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking changes:
 
+- `grammarFromScriptElement` and `grammarsFromScriptElements` have been removed.
 - The ohm-js package now requires an ES2017 environment. This means that very old browsers (e.g., Internet Explorer and Edge 14 and below) are no longer supported. See [caniuse](https://caniuse.com/async-functions,object-values,object-entries,mdn-javascript_builtins_object_getownpropertydescriptors,pad-start-end,mdn-javascript_grammar_trailing_commas_trailing_commas_in_functions) for an overview.
 - There is no longer a default semantic action for iteration nodes. See [here](https://github.com/harc/ohm/blob/master/doc/releases/ohm-js-16.0.md#default-semantic-actions) for more details.
 - (_TypeScript only_): On semantics objects, the methods `addOperation`, `extendOperation`, `addAttribute`, and `extendAttribute` now have a required type parameter, which is the return type of the operation (or the type of the attribute). The related types `Action` and `ActionDict` are similarly parameterized by the return type of the semantic actions. See the [release notes](https://github.com/harc/ohm/blob/master/doc/releases/ohm-js-16.0.md#type-parameters-for-operations-and-actions-typescript-only) for more details.